splitting up how contexts are created so that circulare dependencies won't occur when implementing the CAS validation
