+++ csss-site/src/about/views/officer_management.py
@@ -6,7 +6,8 @@ import markdown
 
 from about.models import Officer, Term
 from about.views.officer_position_and_github_mapping.officer_management_helper import TAB_STRING
-from csss.views_helper import create_main_context, get_current_term, get_previous_term
+from csss.views.context_creation.create_main_context import create_main_context
+from csss.views_helper import get_current_term, get_previous_term
 
 logger = logging.getLogger('csss_site')
