+++ csss-site/src/csss/views/context_creation/create_base_context.py
@@ -1,0 +1,15 @@
+from django.conf import settings
+
+
+def create_base_context():
+    """
+    creates the base context dictionary that contains only the URL_ROOT
+
+    Return
+    context -- the base context dictionary
+
+    """
+    context = {
+        'URL_ROOT': settings.URL_ROOT,
+    }
+    return context
