+++ csss-site/src/elections/views/endpoints/election_page.py
@@ -5,7 +5,8 @@ from django.http import HttpResponseRedirect
 from django.shortcuts import render
 
 from administration.views import user_has_election_management_privilege
-from csss.views_helper import create_main_context, ERROR_MESSAGE_KEY
+from csss.views.context_creation.create_main_context import create_main_context
+from csss.views_helper import ERROR_MESSAGE_KEY
 from elections.models import Election, NomineePosition, NomineeLink
 from elections.views.Constants import TAB_STRING, INPUT_ELECTION_ID__VALUE, ELECTION_MANAGEMENT_PERMISSION, \
     BUTTON_MODIFY_ELECTION_ID__NAME, INPUT_ELECTION_ID__NAME, ELECTION_ID, ELECTION__HTML_NAME, NOMINEES_HTML__NAME, \
