+++ csss-site/src/elections/views/endpoints/json/display_and_process_html_for_json.py
@@ -3,7 +3,8 @@ import logging
 
 from django.shortcuts import render
 
-from csss.views_helper import verify_access_logged_user_and_create_context_for_elections, ERROR_MESSAGE_KEY, \
+from csss.views.request_validation import verify_access_logged_user_and_create_context_for_elections
+from csss.views_helper import ERROR_MESSAGE_KEY, \
     ERROR_MESSAGES_KEY
 from elections.views.Constants import ELECTION_ID, TAB_STRING, UPDATE_EXISTING_ELECTION__NAME
 from elections.views.update_election.json.display_json_for_selected_election_json import \
