+++ csss-site/src/elections/views/endpoints/list_of_elections.py
@@ -2,7 +2,7 @@ import logging
 
 from django.shortcuts import render
 
-from csss.views_helper import create_main_context
+from csss.views.context_creation.create_main_context import create_main_context
 from elections.views.Constants import TAB_STRING
 
 logger = logging.getLogger('csss_site')
