+++ csss-site/src/elections/views/endpoints/nominee_links/display_and_process_html_for_nominee_modification_via_passphrase__nominee_link.py
@@ -3,7 +3,7 @@ import logging
 
 from django.shortcuts import render
 
-from csss.views_helper import create_main_context
+from csss.views.context_creation.create_main_context import create_main_context
 from elections.models import NomineeLink
 from elections.views.Constants import TAB_STRING, CREATE_OR_UPDATE_NOMINEE__NAME, HTML_PASSPHRASE_GET_KEY
 from elections.views.create_context.nominee_links.create_or_update_nominee__nominee_links_html import \
