+++ csss-site/src/events/fall_hacks/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render
 
-from csss.views_helper import create_main_context
+from csss.views.context_creation.create_main_context import create_main_context
 
 TAB = 'events'
