+++ csss-site/src/events/frosh/views/create_frosh_context.py
@@ -1,0 +1,11 @@
+from csss.views.context_creation.create_base_context import create_base_context
+
+
+def create_frosh_context():
+    """
+    creates the context dictionary for the frosh webpages
+
+    Return
+    context -- the frosh webpages context dictionary
+    """
+    return create_base_context()
