+++ csss-site/src/resource_management/views/gdrive_views.py
@@ -6,8 +6,8 @@ from django.http import HttpResponseRedirect
 from django.shortcuts import render
 from querystring_parser import parser
 
-from csss.views_helper import there_are_multiple_entries, verify_access_logged_user_and_create_context, \
-    ERROR_MESSAGE_KEY, ERROR_MESSAGES_KEY
+from csss.views.request_validation import verify_access_logged_user_and_create_context
+from csss.views_helper import there_are_multiple_entries, ERROR_MESSAGE_KEY, ERROR_MESSAGES_KEY
 from resource_management.models import NonOfficerGoogleDriveUser, GoogleDrivePublicFile
 from .get_officer_list import get_list_of_officer_details_from_past_specified_terms
 from .resource_apis.gdrive.gdrive_api import GoogleDrive
